refactor(runner): Rename helper package to execution
Use the execution package for extracted runner-adjacent grading helpers instead of runner_support.
